Implement generator:iconpack kmp module
Delete unused imports in generator:iconpack

Use expect/actual IconPackGenerator object

Revert changes in NewPackViewModel

Simplify assertions to a single equality check

Rename private extension function to clearBlankLines
