+++ components/generator/iconpack/src/commonMain/kotlin/io/github/composegears/valkyrie/generator/iconpack/IconPackGenerator.kt
@@ -1,0 +1,5 @@
+package io.github.composegears.valkyrie.generator.iconpack
+
+expect object IconPackGenerator {
+    fun create(config: IconPackGeneratorConfig): IconPackSpecOutput
+}
