+++ components/generator/iconpack/src/jvmMain/kotlin/io/github/composegears/valkyrie/generator/iconpack/IconPackGenerator.jvm.kt
@@ -1,0 +1,7 @@
+package io.github.composegears.valkyrie.generator.iconpack
+
+actual object IconPackGenerator {
+    actual fun create(config: IconPackGeneratorConfig): IconPackSpecOutput {
+        return IconPackFileSpec(config).createSpec()
+    }
+}
