--- components/generator/jvm/iconpack/src/main/kotlin/io/github/composegears/valkyrie/generator/jvm/iconpack/IconPackGenerator.kt
@@ -1,20 +1,0 @@
-package io.github.composegears.valkyrie.generator.jvm.iconpack
-
-import io.github.composegears.valkyrie.generator.core.IconPack
-
-data class IconPackGeneratorConfig(
-    val packageName: String,
-    val iconPack: IconPack,
-    val useExplicitMode: Boolean,
-    val indentSize: Int,
-)
-
-data class IconPackSpecOutput(
-    val content: String,
-    val name: String,
-)
-
-object IconPackGenerator {
-
-    fun create(config: IconPackGeneratorConfig) = IconPackFileSpec(config).createSpec()
-}
