+++ settings.gradle.kts
@@ -52,8 +52,8 @@ include("tools:idea-plugin")
 
 include("components:extensions")
 include("components:generator:core")
+include("components:generator:iconpack")
 include("components:generator:jvm:poet-extensions")
-include("components:generator:jvm:iconpack")
 include("components:generator:jvm:imagevector")
 include("components:ir")
 include("components:ir-compose")
