+++ tools/idea-plugin/build.gradle.kts
@@ -13,7 +13,7 @@ version = rootProject.providers.gradleProperty("VERSION_NAME").get()
 
 dependencies {
     implementation(projects.components.extensions)
-    implementation(projects.components.generator.jvm.iconpack)
+    implementation(projects.components.generator.iconpack)
     implementation(projects.components.generator.jvm.imagevector)
     implementation(projects.components.irCompose)
     implementation(projects.components.parser.unified)
