+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/mode/iconpack/creation/common/util/IconPackWriter.kt
@@ -4,8 +4,8 @@ import com.intellij.openapi.application.ReadResult.Companion.writeAction
 import com.intellij.openapi.vfs.VirtualFileManager
 import io.github.composegears.valkyrie.extensions.writeToKt
 import io.github.composegears.valkyrie.generator.core.IconPack
-import io.github.composegears.valkyrie.generator.jvm.iconpack.IconPackGenerator
-import io.github.composegears.valkyrie.generator.jvm.iconpack.IconPackGeneratorConfig
+import io.github.composegears.valkyrie.generator.iconpack.IconPackGenerator
+import io.github.composegears.valkyrie.generator.iconpack.IconPackGeneratorConfig
 import io.github.composegears.valkyrie.settings.InMemorySettings
 import io.github.composegears.valkyrie.settings.updateNestedPack
 import io.github.composegears.valkyrie.shared.Mode
