+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/mode/iconpack/existingpack/ui/viewmodel/ExistingPackViewModel.kt
@@ -4,8 +4,8 @@ import com.composegears.tiamat.TiamatViewModel
 import com.intellij.openapi.project.Project
 import io.github.composegears.valkyrie.extensions.safeAs
 import io.github.composegears.valkyrie.generator.core.IconPack
-import io.github.composegears.valkyrie.generator.jvm.iconpack.IconPackGenerator
-import io.github.composegears.valkyrie.generator.jvm.iconpack.IconPackGeneratorConfig
+import io.github.composegears.valkyrie.generator.iconpack.IconPackGenerator
+import io.github.composegears.valkyrie.generator.iconpack.IconPackGeneratorConfig
 import io.github.composegears.valkyrie.psi.iconpack.IconPackInfo
 import io.github.composegears.valkyrie.psi.iconpack.IconPackPsiParser
 import io.github.composegears.valkyrie.ui.di.DI
